Upgrade up-to-date

+++ app/models/attachment.rb
@@ -25,7 +25,6 @@ class Attachment < ActiveRecord::Base
     col = table ? "#{table}.display_name" : 'display_name'
     best_unicode_collation_key(col)
   end
-  strong_params
 
   PERMITTED_ATTRIBUTES = [:filename, :display_name, :locked, :position, :lock_at,
     :unlock_at, :uploaded_data, :hidden, :viewed_at].freeze
@@ -39,6 +38,9 @@ class Attachment < ActiveRecord::Base
   include HasContentTags
   include ContextModuleItem
   include SearchTermHelper
+  include MasterCourses::Restrictor
+  restrict_columns :content, [:display_name, :uploaded_data]
+  restrict_columns :settings, [:folder_id, :locked, :lock_at, :unlock_at, :usage_rights_id]
 
   attr_accessor :podcast_associated_asset
 
@@ -121,7 +123,7 @@ class Attachment < ActiveRecord::Base
   has_attachment(
       :storage => self.store_type.key,
       :path_prefix => file_store_config['path_prefix'],
-      :s3_access => :private,
+      :s3_access => 'private',
       :thumbnails => { :thumb => '128x128' },
       :thumbnail_class => 'Thumbnail'
   )
@@ -251,8 +253,8 @@ class Attachment < ActiveRecord::Base
   attr_accessor :clone_updated
   def clone_for(context, dup=nil, options={})
     if !self.cloned_item && !self.new_record?
-      self.cloned_item ||= ClonedItem.create(:original_item => self)
-      self.save!
+      self.cloned_item = ClonedItem.create(:original_item => self) # do we even use this for anything?
+      Attachment.where(:id => self).update_all(:cloned_item_id => self.cloned_item.id) # don't touch it for no reason
     end
     existing = context.attachments.active.find_by_id(self)
     existing ||= self.cloned_item_id ? context.attachments.active.where(cloned_item_id: self.cloned_item_id).first : nil
@@ -263,19 +265,25 @@ class Attachment < ActiveRecord::Base
 
     excluded_atts = EXCLUDED_COPY_ATTRIBUTES
     excluded_atts += ["locked", "hidden"] if dup == existing
-    dup.assign_attributes(self.attributes.except(*excluded_atts), :without_protection => true)
+    dup.assign_attributes(self.attributes.except(*excluded_atts))
 
-    dup.write_attribute(:filename, self.filename)
     # avoid cycles (a -> b -> a) and self-references (a -> a) in root_attachment_id pointers
     if dup.new_record? || ![self.id, self.root_attachment_id].include?(dup.id)
       dup.root_attachment_id = self.root_attachment_id || self.id
     end
+    dup.write_attribute(:filename, self.filename) unless dup.root_attachment_id?
     dup.context = context
-    dup.migration_id = CC::CCHelper.create_key(self)
+    dup.migration_id = options[:migration_id] || CC::CCHelper.create_key(self)
+    dup.mark_as_importing!(options[:migration]) if options[:migration]
     if context.respond_to?(:log_merge_result)
       context.log_merge_result("File \"#{dup.folder && dup.folder.full_name}/#{dup.display_name}\" created")
     end
-    dup.updated_at = Time.now
+    if Attachment.s3_storage? && context.respond_to?(:root_account_id) && self.namespace != context.root_account.file_namespace
+      dup.save_without_broadcasting!
+      dup.make_rootless
+      dup.change_namespace(context.root_account.file_namespace)
+    end
+    dup.updated_at = Time.zone.now
     dup.clone_updated = true
     dup.set_publish_state_for_usage_rights unless self.locked?
     dup
@@ -315,6 +323,7 @@ class Attachment < ActiveRecord::Base
   def assert_attachment
     if !self.to_be_zipped? && !self.zipping? && !self.errored? && !self.deleted? && (!filename || !content_type || !downloadable?)
       self.errors.add(:base, t('errors.not_found', "File data could not be found"))
+      throw :abort unless CANVAS_RAILS4_2
       return false
     end
   end
@@ -505,7 +514,7 @@ class Attachment < ActiveRecord::Base
           existing_attachment.write_attribute(:filename, nil)
           existing_attachment.save!
           Attachment.where(root_attachment_id: existing_attachment).update_all(
-            root_attachment_id: self,
+            root_attachment_id: id,
             filename: nil,
             updated_at: Time.zone.now)
         end
@@ -643,11 +652,18 @@ class Attachment < ActiveRecord::Base
 
   def handle_duplicates(method, opts = {})
     return [] unless method.present? && self.folder
+
     if self.folder.for_submissions?
       method = :rename
     else
       method = method.to_sym
     end
+
+    if method == :overwrite
+      atts = self.shard.activate { self.folder.active_file_attachments.where("display_name=? AND id<>?", self.display_name, self.id).to_a }
+      method = :rename if atts.any? { |att| att.editing_restricted?(:any) }
+    end
+
     deleted_attachments = []
     if method == :rename
       self.save! unless self.id
@@ -668,15 +684,14 @@ class Attachment < ActiveRecord::Base
           end
         end
       end
-    elsif method == :overwrite
-      atts = self.folder.active_file_attachments.where("display_name=? AND id<>?", self.display_name, self.id)
-      atts.update_all(:replacement_attachment_id => self) # so we can find the new file in content links
-      copy_access_attributes!(atts.first) unless atts.empty?
+    elsif method == :overwrite && atts.any?
+      Attachment.where(:id => atts).update_all(replacement_attachment_id: self.id) # so we can find the new file in content links
+      copy_access_attributes!(atts)
       atts.each do |a|
         # update content tags to refer to the new file
-        ContentTag.where(:content_id => a, :content_type => 'Attachment').update_all(:content_id => self)
+        ContentTag.where(:content_id => a, :content_type => 'Attachment').update_all(content_id: self.id)
         # update replacement pointers pointing at the overwritten file
-        context.attachments.where(:replacement_attachment_id => a).update_all(:replacement_attachment_id => self)
+        context.attachments.where(:replacement_attachment_id => a).update_all(replacement_attachment_id: self.id)
         # delete the overwritten file (unless the caller is queueing them up)
         a.destroy unless opts[:caller_will_destroy]
         deleted_attachments << a
@@ -685,7 +700,9 @@ class Attachment < ActiveRecord::Base
     return deleted_attachments
   end
 
-  def copy_access_attributes!(source)
+  def copy_access_attributes!(source_attachments)
+    self.could_be_locked = true if source_attachments.any?(&:could_be_locked?)
+    source = source_attachments.first
     self.file_state = 'hidden' if source.file_state == 'hidden'
     self.locked = source.locked
     self.unlock_at = source.unlock_at
@@ -888,12 +905,12 @@ class Attachment < ActiveRecord::Base
     h.number_to_human_size(self.size) rescue "size unknown"
   end
 
-  def download_url
-    authenticated_s3_url(:expires => url_ttl, :response_content_disposition => "attachment; " + disposition_filename)
+  def download_url(ttl = url_ttl)
+    authenticated_s3_url(expires_in: ttl, response_content_disposition: "attachment; " + disposition_filename)
   end
 
-  def inline_url
-    authenticated_s3_url(:expires => url_ttl, :response_content_disposition => "inline; " + disposition_filename)
+  def inline_url(ttl = url_ttl)
+    authenticated_s3_url(expires_in: ttl, response_content_disposition: "inline; " + disposition_filename)
   end
 
   def url_ttl
@@ -932,10 +949,9 @@ class Attachment < ActiveRecord::Base
     super(name)
   end
 
-  def thumbnail_with_root_attachment
-    self.thumbnail_without_root_attachment || self.root_attachment.try(:thumbnail)
+  def thumbnail
+    super || root_attachment.try(:thumbnail)
   end
-  alias_method_chain :thumbnail, :root_attachment
 
   def content_directory
     self.directory_name || Folder.root_folders(self.context).first.name
@@ -1212,23 +1228,63 @@ class Attachment < ActiveRecord::Base
     true
   end
 
+  # this will delete the content of the attachment but not delete the attachment
+  # object. It will replace the attachment content with a file_removed file.
+  def destroy_content_and_replace
+    raise 'must be a root_attachment' if self.root_attachment_id
+    self.destroy_content
+    self.uploaded_data = File.open Rails.root.join('public/file_removed/file_removed.pdf')
+    CrocodocDocument.where(attachment_id: self).delete_all
+    Canvadoc.where(attachment_id: self).delete_all
+    self.save!
+  end
+
+  def dmca_file_removal
+    destroy_content_and_replace
+  end
+
+  def destroy_content
+    raise 'must be a root_attachment' if self.root_attachment_id
+    return unless self.filename
+    if Attachment.s3_storage?
+      self.s3object.delete
+    else
+      FileUtils.rm full_filename
+    end
+  end
+
   def make_childless(preferred_child = nil)
+    return if root_attachment_id
     child = preferred_child || children.take
+    return unless child
     raise "must be a child" unless child.root_attachment_id == id
     child.root_attachment_id = nil
-    child.filename = filename if filename
+    copy_attachment_content(child)
+    Attachment.where(root_attachment_id: self).where.not(id: child).update_all(root_attachment_id: child.id)
+  end
+
+  def copy_attachment_content(destination)
+    destination.write_attribute(:filename, filename) if filename
     if Attachment.s3_storage?
-      if filename && s3object.exists? && !child.s3object.exists?
-        s3object.copy_to(child.s3object)
+      if filename && s3object.exists? && !destination.s3object.exists?
+        s3object.copy_to(destination.s3object)
       end
     else
+      return if destination.store.exists? && open == destination.open
       old_content_type = self.content_type
       Attachment.where(:id => self).update_all(:content_type => "invalid/invalid") # prevents find_existing_attachment_for_md5 from reattaching the child to the old root
-      child.uploaded_data = open
+      destination.uploaded_data = open
       Attachment.where(:id => self).update_all(:content_type => old_content_type)
     end
-    child.save!
-    Attachment.where(root_attachment_id: self).where.not(id: child).update_all(root_attachment_id: child)
+    destination.save!
+  end
+
+  def make_rootless
+    return unless root_attachment_id
+    root = self.root_attachment
+    return unless root
+    self.root_attachment_id = nil
+    root.copy_attachment_content(self)
   end
 
   def restore
@@ -1590,7 +1646,12 @@ class Attachment < ActiveRecord::Base
         attachment.filename ||= File.basename(uri.path)
         attachment.uploaded_data = tmpfile
         if attachment.content_type.blank? || attachment.content_type == "unknown/unknown"
-          attachment.content_type = http_response.content_type
+          # uploaded_data= clobbers the content_type set in preflight; if it was given, prefer it to the HTTP response
+          attachment.content_type = if attachment.content_type_was.present? && attachment.content_type_was != 'unknown/unknown'
+            attachment.content_type_was
+          else
+            http_response.content_type
+          end
         end
         return attachment
       else
@@ -1623,7 +1684,7 @@ class Attachment < ActiveRecord::Base
           end
         else
           new_attachment = Attachment.new
-          new_attachment.assign_attributes(attachment.attributes.except(*EXCLUDED_COPY_ATTRIBUTES), :without_protection => true)
+          new_attachment.assign_attributes(attachment.attributes.except(*EXCLUDED_COPY_ATTRIBUTES))
 
           new_attachment.user_id = to_context.id if to_context.is_a? User
           new_attachment.context = to_context
